requirements.txt
Flask==1.1.1
gevent==1.4.0
gunicorn==19.9.0
glob2==0.7
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Theano==1.0.4
tensorflow>=1.15.2
Werkzeug==2.2.3
glob2==0.7
future==0.18.3
numpy==1.22.0
streamlit
tensorflow
pillow
numpy

